standards(ruff): clear ruff lint across tests/, editor/, tools/, main.py
Sonnet fan-out over 5 disjoint path-groups, 264 findings -> 0. With this `ruff
check .` is GREEN repo-wide (the test_ruff_lint gate passes) and ruff format is
clean (315 files).

Behavior-preserving lint fixes only — for test files, NO assertion was weakened,
narrowed, skipped, or xfailed (the agents ran each edited test file; tripwire
holds at 2657 passed / 0 failed). Breakdown: RUF059 unused-unpack -> _name;
F841 dead locals removed (call side-effects kept); E741/E712/B905/B904/SIM105/
ERA001/RUF046/RUF005/RUF007/RUF012/PERF401 per recipe; a few specific
`# noqa: CODE + reason` where the lint would have changed intent.

Notable: main.py build_demo C901 (complexity 43) reduced to 12 by extracting its
11 inner key-binding callbacks to module-level helpers bound via functools.partial
(+ _build_gpu_pipeline / _log_renderer_info / _do_boot_load). Behavior verified:
tools/screenshot.py --explode booted the full demo, fired the explosion callback
("2 chunks touched"), rendered 120 frames, captured a valid framebuffer, exit 0.

+++ tests/test_wind_worker_edges.py
@@ -65,7 +65,7 @@ def _solid_job(cfg: Config, *, seq: int = 1, cells: int = 8) -> VenturiJob:
     """
     ground = float(cfg.ground_height_m)
     cell_m = float(cfg.wind_cell_m)
-    vpc = int(round(cell_m / VOXEL))
+    vpc = round(cell_m / VOXEL)
     region_v = cells * vpc
     vz_lo = int(np.floor(ground / VOXEL))
     vz_hi = int(np.ceil((ground + float(cfg.wind_layer_m)) / VOXEL))
